% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_exceltable.R
\name{absfilepath}
\alias{absfilepath}
\title{Find avbsolute path of file}
\usage{
absfilepath(relativepath)
}
\arguments{
\item{relativepath}{A relative path to a file}
}
\value{
the absolute path to a file
}
\description{
Find avbsolute path of file
}
\examples{

\dontrun{
absfilepath("../path/to/file")
}






}
