% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_normfinder.R
\name{adjustnmfRounding}
\alias{adjustnmfRounding}
\title{Adjust rounding in the normfinder algorithm to 3 decimals}
\usage{
adjustnmfRounding(decimals = 3)
}
\arguments{
\item{decimals}{An integer value indicating desired precision rounding of the Normfinder results}
}
\value{
Void
}
\description{
Adjust rounding in the normfinder algorithm to 3 decimals
}
