% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refSeeker-package.R
\docType{package}
\name{refSeeker-package}
\alias{refSeeker}
\alias{refSeeker-package}
\title{refSeeker: A RefFinder stability ranking tool}
\description{
Simple calculation and extraction of expression stablity and stablity ranking based on the RefFinder method.
}
\keyword{internal}
