% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_excelbatch.R
\name{rs_batchExcel}
\alias{rs_batchExcel}
\title{Creates batch output from selected excel file}
\usage{
rs_batchExcel()
}
\value{
Nothing. Creates excel files and graphs with stabilities for each sheet of a selected excel file
}
\description{
Creates batch output from selected excel file
}
\examples{
\dontrun{
rs_batchExcel()
}



}
