% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_excelbatch.R
\name{rf_batchExcel2}
\alias{rf_batchExcel2}
\title{Creates batch output from selected excel file}
\usage{
rf_batchExcel2(input = "", outdir = "", printgraph = TRUE)
}
\arguments{
\item{input}{A path to an excel file containing properly formatted reffinder data tables. This means column representing genes/targets and rows representing samples,
data being Ct/Cp/Cq values, with no missing data since the validity of these results have not been confirmed.}

\item{outdir}{A selected output directory, if not set a dialog will appear to allow selection for output directory}

\item{printgraph}{If TRUE print a graph of the result in the output folder}
}
\value{
Nothing. Creates excel files and graphs with stabilities for each sheet of a selected excel file
}
\description{
Creates batch output from selected excel file
}
\examples{
\dontrun{
rf_batchExcel()
}




}
