% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_bestKeeper.R
\name{rs_bestkeeper}
\alias{rs_bestkeeper}
\title{Calculates stabilites based on the BestKeeper method}
\usage{
rs_bestkeeper(expression, decimals = 3)
}
\arguments{
\item{expression}{A matrix, data frame or tibble with columns for each genes and rows for each samples.}

\item{decimals}{An integer value indicating result precision. Used for the rounding of results.}
}
\value{
A data frame with stability values - std dev +/- CP
}
\description{
Calculates stabilites based on the BestKeeper method
}
\examples{
set.seed(100)
ct_vals <- data.frame(matrix(rnorm(5*20, mean = 25), ncol = 5, nrow = 20))

names(ct_vals) <- c("gene1", "gene2", "gene3", "gene4", "gene5")
bks <- rs_bestkeeper(ct_vals)
bks

}
\references{
Pfaffl MW, Tichopad A, Prgomet C, Neuvians TP. 2004. Determination of stable housekeeping genes, differentially regulated target genes and sample integrity: BestKeeper--Excel-based tool using pair-wise correlations. Biotechnology letters 26:509-515.
}
