% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_csvbatch.R
\name{rs_csvbatch}
\alias{rs_csvbatch}
\title{Creates batch output from selected comma separated txt files}
\usage{
rs_csvbatch(input = "", outdir = "", outtype = "csv", printgraph = TRUE)
}
\arguments{
\item{input}{A path to a file txt-like files (txt, csv, tsv etc.) containing a properly formatted reffinder data table. All files with the same file extension in that folder will be loaded as a data set. (See vignette about preparing and loading txt-like files)}

\item{outdir}{A selected output directory, if not set a dialog will appear to allow selection for output directory}

\item{outtype}{type of data output can be "csv", "excel" or "tsv" - Have not been implemented yet}

\item{printgraph}{If TRUE print a graph of the result in the output folder}
}
\value{
Nothing. Creates excel files and graphs with stabilities for each sheet of a selected excel file
}
\description{
Creates batch output from selected comma separated txt files
}
\examples{
\dontrun{
rs_csvbatch()
}



}
