% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_deltaCt.R
\name{rs_deltaCt}
\alias{rs_deltaCt}
\title{Calculates expression stabilities by the Delta-Ct method}
\usage{
rs_deltaCt(expression)
}
\arguments{
\item{expression}{A matrix, data frame or tibble with columns for each genes and rows for each samples.}
}
\value{
Data frame with mean standard deviations for pairwise delta Ct values
}
\description{
Calculates expression stabilities by the Delta-Ct method
}
\examples{
set.seed(100)
ct_vals <- matrix(rnorm(5*20, mean = 25), ncol = 5, nrow = 20)
dimnames(ct_vals)[[2]] <-  c("gene1", "gene2", "gene3", "gene4", "gene5")

dct <- rs_deltaCt(ct_vals)
dct

########################

set.seed(100)
ct_vals <- data.frame(matrix(rnorm(5*20, mean = 25), ncol = 5, nrow = 20))

names(ct_vals) <- c("gene1", "gene2", "gene3", "gene4", "gene5")
dct <- rs_deltaCt(ct_vals)
dct

}
\references{
Silver N, Best S, Jiang J, Thein SL. 2006. Selection of housekeeping genes for gene expression studies in human reticulocytes using real-time PCR. BMC molecular biology 7:33.
}
