% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_exceltable.R
\name{rf_exceltable}
\alias{rf_exceltable}
\title{Output excel workbook containing stability and ranking as separate sheets}
\usage{
rf_exceltable(reffinderlist, filename = "Stability-table", addDate = TRUE)
}
\arguments{
\item{reffinderlist}{A reffinderlist created by the reffinder function}

\item{filename}{A file name to identify the data}

\item{addDate}{Possibility to automatically add date to output file name}
}
\value{
Nothing, outpust an excel file in the working directory or an output folder selected by the file name
}
\description{
Output excel workbook containing stability and ranking as separate sheets
}
\examples{
set.seed(100)
ct_vals <- matrix(rnorm(5*20, mean = 25), ncol = 5, nrow = 20)
dimnames(ct_vals)[[2]] <-  c("gene1", "gene2", "gene3", "gene4", "gene5")

\dontrun{
rffndr <- rf_reffinder(ct_vals)

rf_exceltable(rffndr)
}


}
