% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_exporttables.R
\name{rs_exporttable}
\alias{rs_exporttable}
\title{Export RefSeeker results tables}
\usage{
rs_exporttable(
  refseekerlist,
  filename = "Stability-table",
  tabletype = "xlsx",
  addDate = TRUE
)
}
\arguments{
\item{refseekerlist}{A RefSeeker list normally provided by rs_reffinder() function}

\item{filename}{A filename for the output, may include relative path to desired folder}

\item{tabletype}{Select type of output may be one of ; "xlsx", "ods, csv", "tsv", "txt",  "docx-stability",  "docx-rank" or  "docx-combi"}

\item{addDate}{Should the current date be added to the file name}
}
\value{
Nothing but creates files containg table data
}
\description{
Export RefSeeker results tables
}
\examples{
set.seed(100)
ct_vals <- matrix(rnorm(5*20, mean = 25), ncol = 5, nrow = 20)
dimnames(ct_vals)[[2]] <-  c("gene1", "gene2", "gene3", "gene4", "gene5")
res <- rs_reffinder(ct_vals)

\dontrun{
rs_exporttable(res, filename = "../data/EOCstability", tabletype = "csv", addDate = TRUE )

}



}
