% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_genorm.R
\name{rf_genorm}
\alias{rf_genorm}
\title{Gene expression stability by geNorm}
\usage{
rf_genorm(expression)
}
\arguments{
\item{expression}{A matrix, data frame or tibble with columns for each genes and rows for each samples.}
}
\value{
Data frame with average pairwise variation for each gene
}
\description{
Gene expression stability by geNorm
}
\examples{
set.seed(100)
ct_vals <- matrix(rnorm(5*20), ncol = 5, nrow = 20)
dimnames(ct_vals)[[2]] <-  c("gene1", "gene2", "gene3", "gene4", "gene5")

gnm <- rf_genorm(ct_vals)
gnm

########################

set.seed(100)
ct_vals <- data.frame(matrix(rnorm(5*20, mean = 25), ncol = 5, nrow = 20))

names(ct_vals) <- c("gene1", "gene2", "gene3", "gene4", "gene5")
gnm <- rf_genorm(ct_vals)
gnm


}
\references{
Vandesompele J, De Preter K, Pattyn F, Poppe B, Van Roy N, De Paepe A, Speleman F. 2002. Accurate normalization of real-time quantitative RT-PCR data by geometric averaging of multiple internal control genes. Genome biology 3:RESEARCH0034.
}
