% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_loadTextBaasedTables.R
\name{rs_load.table}
\alias{rs_load.table}
\title{Read in expresion data files with auto delimiter detection}
\usage{
rs_load.table(files = "")
}
\arguments{
\item{files}{An optional file path or a vector of file paths one or more data files}
}
\value{
A list or list of lists of input data tables
}
\description{
Read in expresion data files with auto delimiter detection
}
\examples{

\dontrun{
rs_loadcsvdata()
}



}
