% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_loadexceldata.R
\name{rs_loaddata}
\alias{rs_loaddata}
\title{Load data from a supported source}
\usage{
rs_loaddata(files = "")
}
\arguments{
\item{files}{A character value or vector containing path(s) for files to be loaded into R. If left empty a dialog will allow for selection of files}
}
\value{
a list of data frames contain the selected data sets
}
\description{
Load data from a supported source
}
\examples{
\dontrun{
rs_loaddata()
}
}
