% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_loadexceldata.R
\name{rf_loadexceldata}
\alias{rf_loadexceldata}
\title{Load one or more datasets from an excel file}
\usage{
rf_loadexceldata(filepath = "")
}
\arguments{
\item{filepath}{a path to output the data files}
}
\value{
A list containing expression data for each sheet in the excel file
}
\description{
Load one or more datasets from an excel file
}
\examples{

\dontrun{
rf_loadexceldata()
}


}
