% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_miRNA.R
\docType{data}
\name{rf_miRNA}
\alias{rf_miRNA}
\title{miRNA expression data from ovarian cancer patients}
\format{
A tibble containing miRNA Cp values using miRNA ID as columns and samples as rows
}
\usage{
data(rf_miRNA)
}
\description{
miRNA expression data from ovarian cancer patients
}
\examples{

data(rf_miRNA)

\dontrun{
rf_reffinder(rf_miRNA)
}
}
\keyword{datasets}
