% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_normfinder.R
\name{rs_normfinder}
\alias{rs_normfinder}
\title{Gene expresssion stabilities by NormFinder}
\usage{
rs_normfinder(expression)
}
\arguments{
\item{expression}{A matrix, data frame or tibble with columns for each genes and rows for each samples.}
}
\value{
Data frame with stability values based on the Normfinder method
}
\description{
Gene expresssion stabilities by NormFinder
}
\examples{
set.seed(100)
ct_vals <- matrix(rnorm(5*20, mean = 25), ncol = 5, nrow = 20)
dimnames(ct_vals)[[2]] <-  c("gene1", "gene2", "gene3", "gene4", "gene5")
\dontrun{

nrmfndr <- rs_normfinder(ct_vals)
nrmfndr

}

########################

set.seed(100)
ct_vals <- data.frame(matrix(rnorm(5*20, mean = 25), ncol = 5, nrow = 20))

names(ct_vals) <- c("gene1", "gene2", "gene3", "gene4", "gene5")

\dontrun{
nrmfndr <- rs_normfinder(ct_vals)
nrmfndr
}



}
\references{
Andersen C.L., Ledet-Jensen J., Ørntoft T.: Normalization of real-time quantitative RT-PCR data: a model based variance estimation approach to identify genes suited for normalization - applied to bladder- and colon-cancer data-sets.
Cancer Research. 2004 (64): 5245-5250
}
