% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_normfinder.R
\name{rs_normfinderFull}
\alias{rs_normfinderFull}
\title{Title}
\usage{
rs_normfinderFull(expression, Groups = TRUE, ctVal = TRUE, pStabLim = 0.25)
}
\arguments{
\item{expression}{A matrix, data frame or tibble with columns for each genes and rows for each samples.}

\item{Groups}{Does the data consist of different treatment groups, if so indicate group ID in the last column of the expression data set}

\item{ctVal}{Is the provided expression data ca values or relative log(2) transformed values?}

\item{pStabLim}{Stability limit, report only stabilities above this value}
}
\value{
A list of lists of result tables depending on the Groups parameter
}
\description{
Title
}
\references{
Andersen C.L., Ledet-Jensen J., Ørntoft T.: Normalization of real-time quantitative RT-PCR data: a model based variance estimation approach to identify genes suited for normalization - applied to bladder- and colon-cancer data-sets.
Cancer Research. 2004 (64): 5245-5250
}
