% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswizzard.R
\name{rs_wizzard}
\alias{rs_wizzard}
\title{Perform stability analysis through interactive wizzard}
\usage{
rs_wizzard()
}
\value{
Creates graphs and result export to
}
\description{
Perform stability analysis through interactive wizzard
}
\examples{

\dontrun{
rs_wizzard()
}
}
