% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rforderbystability.R
\name{rfaddstabilityrank}
\alias{rfaddstabilityrank}
\title{Adds a ranking colum to at stability table. If two targets have equal stabilities they are given the same rank.}
\usage{
rfaddstabilityrank(stabilitytable, stabilitycol)
}
\arguments{
\item{stabilitytable}{A stability table, normally generated by an rf algorithm}

\item{stabilitycol}{A number indicating the column holding the stability values}
}
\value{
A stability table with ranking column
}
\description{
Adds a ranking colum to at stability table. If two targets have equal stabilities they are given the same rank.
}
