% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_exporttables.R
\name{rscsvtable}
\alias{rscsvtable}
\title{Creates two csv files containing stabiltiy and rankings using comma as separator}
\usage{
rscsvtable(refseekerlist, filename, addDate = TRUE)
}
\arguments{
\item{refseekerlist}{A RefSeekerlist created by the rs_reffinder function}

\item{filename}{A file name to identify the data}

\item{addDate}{Possibility to automatically add date to output file name}
}
\value{
Nothing, output an excel file in the working directory or an output folder selected by the file name

Nothing. Creates csv files in selected directory
}
\description{
Creates two csv files containing stabiltiy and rankings using comma as separator
}
\examples{
\dontrun{

 data <- rs_miRNA
 res <- rs_reffinder(data)
 rscsvtable(res, "path/to/folder", filenameprefix = "EOC")
}





}
