% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_loaddata.R
\name{rsdatatest}
\alias{rsdatatest}
\title{Tests imported data for correct format}
\usage{
rsdatatest(expression)
}
\arguments{
\item{expression}{Imported list of expression data sets}
}
\value{
void outputs warnings if discrepancies are found
}
\description{
Tests imported data for correct format
}
\note{
This is a simple test for missing data and character columns, correctness of data format is not guarantied if no warnings are given
}
