% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_exporttables.R
\name{rsdoctable}
\alias{rsdoctable}
\title{Creates kable tables for refseeker results}
\usage{
rsdoctable(
  refseekerlist,
  filename = "",
  tabletype = "stability",
  addDate = "TRUE"
)
}
\arguments{
\item{refseekerlist}{A RefSeekerlist created by the rs_reffinder function.}

\item{filename}{A file name prefix. May contain a path to an output directory.}

\item{tabletype}{A string representing the table to be printer can be; "stability", "rank" or "both".}

\item{addDate}{Logical indicating whether or not to add the current date to the output file name.}
}
\value{
A kable object in the form of html code string.
}
\description{
Creates kable tables for refseeker results
}
\examples{

\dontrun{
  table <- rsdoctable(results, filename = "~/Desktop/test/doc2_", tabletype = "both")
}


}
