% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_exporttables.R
\name{rsdoctable1}
\alias{rsdoctable1}
\title{Creates tables based on aither stability or rank from a refseeker list.}
\usage{
rsdoctable1(
  refseekerlist,
  filename = "",
  caption = "",
  type = "stability",
  addDate = TRUE
)
}
\arguments{
\item{refseekerlist}{A RefSeekerlist created by the rs_reffinder function}

\item{filename}{A file name prefix. May contain a path to an output directory.}

\item{caption}{A string representing the header of the table.}

\item{type}{A string determining which table to format can be; "stability" or "RankW"}

\item{addDate}{Logical indicating whether or not to add the current date to the output file name.}
}
\value{
A kable object consisting of a html table of choice
}
\description{
Creates tables based on aither stability or rank from a refseeker list.
}
\examples{
\dontrun{
  rsdoctable1(results, filename = "Test1", caption = "Group1", type = "rank")
}



}
