% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_exporttables.R
\name{rsdoctable2}
\alias{rsdoctable2}
\title{Creates tables based on aither stability and rank from a RefSeeker list.}
\usage{
rsdoctable2(refseekerlist, filename = "", caption = "", addDate = TRUE)
}
\arguments{
\item{refseekerlist}{A RefSeeker list created by the rs_reffinder function}

\item{filename}{A file name prefix. May contain a path to an output directory.}

\item{caption}{A string representing the header of the table.}

\item{addDate}{Logical indicating whether or not to add the current date to the output file name.}
}
\value{
A kable object consisting of a html code for a table showing rank and stabiltiy of each reffinder algorithm
}
\description{
Creates tables based on aither stability and rank from a RefSeeker list.
}
\examples{
\dontrun{
  rsdoctable(results, filename = "Test1", caption = "Group1")
}



}
