% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsexcelfile.R
\name{rsexcelfile}
\alias{rsexcelfile}
\title{Creates an excel workbook in the working directory containing the provided tables as sheets}
\usage{
rsexcelfile(filename, ..., addFilter = TRUE, overwrite = FALSE)
}
\arguments{
\item{filename}{An excel file name or path}

\item{...}{One or more tables to be written as sheets}

\item{addFilter}{Toggle filter selection. Adds a filter from first row of table.}

\item{overwrite}{Toggle whether to overwrite a possible existing file}
}
\value{
nothing
}
\description{
Creates an excel workbook in the working directory containing the provided tables as sheets
}
\examples{

set.seed(100)
table1 <- data.frame(matrix(rnorm(5*20, mean = 25), ncol = 5, nrow = 20))

set.seed(200)
table2 <- data.frame(matrix(rnorm(5*20, mean = 25), ncol = 5, nrow = 20))


rsexcelfile("/Output/fileNameofExcelFile", table1, table2, overwrite = TRUE)



}
