% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfoutdirselect.R
\name{rfexltypeselect}
\alias{rfexltypeselect}
\title{Dialog window to select an excel file type}
\usage{
rfexltypeselect()
}
\value{
A string representing the selected type
}
\description{
Dialog window to select an excel file type
}
\examples{

\dontrun{
path <- rfoutdirselect()
}


}
