% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_graph.R
\name{rsgraphdraw}
\alias{rsgraphdraw}
\title{Multifaceted bar graph of stability values from multiple data sets}
\usage{
rsgraphdraw(
  refseekerlist,
  filename = "",
  filetype = "png",
  width,
  height,
  units = "px",
  res = 250,
  ordering = "Comprehensive Rank",
  colors = c("")
)
}
\arguments{
\item{refseekerlist}{A list of data frames from multiple data sets generated by the reffinder function}

\item{filename}{A file name to identify the data. Do not need file extension. Can contain a path if working directory is not the desired destination.}

\item{filetype}{Selection of image file type may be "png", "tiff", "jpeg".}

\item{width}{Parameter passed to the PNG graphics device. Width of the output png. True width is dependent on the selected units-type}

\item{height}{Parameter passed to the PNG graphics device. Height of the output png. True Height is dependent on the selected units-type}

\item{units}{Parameter passed to the PNG graphics device. The units in which height and width are given. Can be "px" (pixels, the default), "in" (inches), "cm" or "mm".}

\item{res}{Parameter passed to the PNG graphics device. The nominal resolution in ppi which will be recorded in the bitmap file, if a positive integer. Also used for units other than the default, and to convert points to pixels.}

\item{ordering}{Used to control sorting of the x axis, use "Target","Delta-Ct","BestKeeper","Normfinder","geNorm" or "Comprehensive Rank".}

\item{colors}{A data frame mapping target names to a hex color with targets in column 1 and colors in columns 2
 target1       "#2271b2"
 target2       "#d55e00"
Gene Delta Ct Bestkeeper Normfinder Genorm Comprehensive Rank}
}
\value{
Creates a png file with the bar graph and put it in an Output folder in the working directory
}
\description{
Multifaceted bar graph of stability values from multiple data sets
}
\examples{
set.seed(100)
ct_vals <- matrix(rnorm(5*20, mean = 25), ncol = 5, nrow = 20)
dimnames(ct_vals)[[2]] <-  c("gene1", "gene2", "gene3", "gene4", "gene5")
\dontrun{
rffndr <- rs_reffinder(ct_vals)

rs_graphdraw(rffndr)
}


###
data(tble2)
\dontrun{
rs_graph(rs_reffinder(tble2))
}
}
