% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfoutdirselect.R
\name{rfgraphtypeselect}
\alias{rfgraphtypeselect}
\title{Dialog window to select a graph type}
\usage{
rfgraphtypeselect()
}
\value{
A string representing the selected type
}
\description{
Dialog window to select a graph type
}
\examples{

\dontrun{
path <- rfoutdirselect()
}


}
