% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_loadexceldata.R
\name{rsloadexceldata}
\alias{rsloadexceldata}
\alias{rsloadodsdata()}
\alias{rsloadodsdata}
\title{Load one or more datasets from an excel file}
\usage{
rsloadexceldata(filepath = "")

rsloadodsdata(filepath = "")
}
\arguments{
\item{filepath}{a path to an excel or ods file. If left blank, a dialog will appear to select this interactivly.}
}
\value{
A list containing expression data for each sheet in the excel file
}
\description{
Load one or more datasets from an excel file
}
\examples{

\dontrun{
rsloadexceldata()
}


\dontrun{
rsloadodsdata()
}
}
