% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_loaddata.R
\name{rsloadtable}
\alias{rsloadtable}
\title{Load data from text based data files.}
\usage{
rsloadtable(files = "")
}
\arguments{
\item{files}{An optional file path or a vector of file paths one or more data files.
If no files are given a dialog will allow selection of files}
}
\value{
A list or list of lists of input data tables
}
\description{
Read in expression data from text based files (csv, tsv, txt, etc.) with auto delimiter detection.
Please note that though the most common uses of the file types has been tested, some delimiter and file type combination may be incompatible.
}
\examples{

\dontrun{

rsloadtable()

rsloadtable("filepath")

rsloadtable(c("fileone", "filetwo"))

}


}
