% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsorderbystability.R
\name{rsorderbystability}
\alias{rsorderbystability}
\title{Orders a expression stability data frame by stability value}
\usage{
rsorderbystability(df)
}
\arguments{
\item{df}{An expression stability matrix, tibble or data frame, should have expression stabilities in second column}
}
\value{
An ordered data frame
}
\description{
Orders a expression stability data frame by stability value
}
\examples{
ct_vals <- matrix(rnorm(5*20), ncol = 5, nrow = 20)
dimnames(ct_vals)[[2]] <-  c("gene1", "gene2", "gene3", "gene4", "gene5")

dct <- rs_deltaCt(ct_vals)

dct <- rsorderbystability(dct)
dct

}
