% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfoutdirselect.R
\name{rfoutdirselect}
\alias{rfoutdirselect}
\title{Dialog window to select a folder for output files}
\usage{
rfoutdirselect()
}
\value{
A string representing a path to a selected output folder
}
\description{
Dialog window to select a folder for output files
}
\examples{

\dontrun{
path <- rfoutdirselect()
}


}
