% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsdialogs.R
\name{rsselectfiles}
\alias{rsselectfiles}
\title{File selection dialog for RefSeeker}
\usage{
rsselectfiles()
}
\value{
A character vector containing the selected files
}
\description{
File selection dialog for RefSeeker
}
\examples{

\dontrun{

rsselectfiles()
}



}
