% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_exporttables.R
\name{rstsvtable}
\alias{rstsvtable}
\title{Creates two tsv files containing stabiltiy and rankings using tab as separator}
\usage{
rstsvtable(refseekerlist, filename, addDate = TRUE)
}
\arguments{
\item{refseekerlist}{A RefSeekerlist created by the rs_reffinder function}

\item{filename}{A file name to identify the data}

\item{addDate}{Possibility to automatically add date to output file name}
}
\value{
Nothing but creates tsv files in selected directory
}
\description{
Creates two tsv files containing stabiltiy and rankings using tab as separator
}
\examples{
\dontrun{

 data <- rs_miRNA
 res <- rs_reffinder(data)
 rstsvtable(res, "path/to/folder", filenameprefix = "EOC")
}


}
