% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_exceltable.R
\name{rstxttable}
\alias{rstxttable}
\title{Creates two txt files containing stabiltiy and rankings using space as separator}
\usage{
rstxttable(refseekerlist, filename, addDate = TRUE)
}
\arguments{
\item{refseekerlist}{A RefSeekerlist created by the rs_reffinder function}

\item{filename}{A file name to identify the data}

\item{addDate}{Possibility to automatically add date to output file name}
}
\value{
Nothing, output an excel file in the working directory or an output folder selected by the file name

Nothing but creates csv files in selected directory
}
\description{
Creates two txt files containing stabiltiy and rankings using space as separator
}
\examples{
\dontrun{

 data <- rs_miRNA
 res <- rs_reffinder(data)
 rstxttable(res, "path/to/folder", filenameprefix = "EOC")
}


}
