% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tble2.R
\docType{data}
\name{tble2}
\alias{tble2}
\title{Expression test data extracted from original RefFinder web tool.}
\format{
A tibble containing miRNA Cp values using miRNA ID as columns and samples as rows
}
\usage{
data(tble2)
}
\description{
Expression test data extracted from original RefFinder web tool.
}
\note{
Data was found online on 2022-06-27
(\href{https://www.ciidirsinaloa.com.mx/RefFinder-master/?type=reference}{RefFinder})

The web tool is provided by Fuliang Xie at github
(\href{https://github.com/fulxie/RefFinder}{github})
}
\examples{
data(tble2)
\dontrun{
rf_reffinder(tble2)
}

}
\references{
F Xie, P Xiao, D Chen, L Xu, B Zhang. 2012. miRDeepFinder: a miRNA analysis tool for deep sequencing of plant small RNAs. Plant molecular biology 80 (1), 75-84.
}
\keyword{datasets}
