% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstestfunctions.R
\name{testrefpath}
\alias{testrefpath}
\title{find the path to the testref folder containing reffences for checking results during testing}
\usage{
testrefpath()
}
\value{
A path to testref folder
}
\description{
find the path to the testref folder containing reffences for checking results during testing
}
