% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refSeeker-package.R
\docType{package}
\name{RefSeeker-package}
\alias{RefSeeker}
\alias{RefSeeker-package}
\title{RefSeeker: A RefFinder stability ranking tool}
\description{
Simple calculation and extraction of expression stability and stability ranking based on the RefFinder method.
}
\keyword{internal}
