% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_excelbatch.R
\name{rf_batchExcel}
\alias{rf_batchExcel}
\title{Creates batch output from selected excel file}
\usage{
rf_batchExcel()
}
\value{
Nothing. Creates excel files and graphs with stabilities for each sheet of a selected excel file
}
\description{
Creates batch output from selected excel file
}
\examples{
\dontrun{
rf_batchExcel()
}




}
