% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_ctrlGene.R
\name{rf_bestKeeperFull}
\alias{rf_bestKeeperFull}
\title{Determines stability of genes}
\usage{
rf_bestKeeperFull(expression, ctVal = TRUE)
}
\arguments{
\item{expression}{A matrix of expression levels. Each row corresponds to a sample and each column to a gene.}

\item{ctVal}{A logical value indicating data type. If ct-values are input, ctVal=TRUE, otherwise, ctVal=FALSE.}
}
\value{
A list containing CP.statistics, pair.Wise.cor and HKG.vs.BestKeeper, which are returned by cpSta(), pearsonCor() and bki(), respectively.
}
\description{
This function is a wrapper calling the bestKeeper function from the ctrlGene package which combines the results of rf_cpSta(), pearsonCor() and bki().
}
\note{
This function was originally published by Shanliang Zhong as part of th ctrlGene package.
}
\examples{

data(rf_miRNA)


rf_bestKeeperFull(as.data.frame(rf_miRNA))
rf_bestKeeperFull(rf_miRNA)


}
\references{
Pfaffl MW, Tichopad A, Prgomet C, Neuvians TP. Biotechnol Lett (2004) <doi: 10.1023/B:BILE.0000019559.84305.47>
}
