% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_loadTextBaasedTables.R
\name{rs_loadcsvdata}
\alias{rs_loadcsvdata}
\title{Read in csv files for processing}
\usage{
rs_loadcsvdata(filepath = "")
}
\arguments{
\item{filepath}{An optional file path to a data file. The function will try to find all files with same extension to load.}
}
\value{
A list or list of lists of input data tables
}
\description{
Read in csv files for processing
}
\note{
This functions has been depricated and will be removed in the near future. Please use rs_load.table() for importing data sets from text based formats"
}
