% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_loadcsvdata.R
\name{rf_loadcsvdata}
\alias{rf_loadcsvdata}
\title{Read in csv files for processing}
\usage{
rf_loadcsvdata(filepath = "")
}
\arguments{
\item{filepath}{An optional file path to a data file. The function try to find all files with same extension to load.}
}
\value{
A list or list of lists of input data tables
}
\description{
Read in csv files for processing
}
\examples{

\dontrun{
rf_loadcsvdata()
}



}
