% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_loadTextBasedTables.R
\name{rs_loadtxtdata}
\alias{rs_loadtxtdata}
\title{Read in reffinder data files with auto delimiter detection}
\usage{
rs_loadtxtdata(filepath = "")
}
\arguments{
\item{filepath}{An optional file path to a data file. The function will try to find all files with same extension to load.}
}
\value{
A list or list of lists of input data tables
}
\description{
Read in reffinder data files with auto delimiter detection
}
\note{
This functions has been depricated and will be removed in the near future. Please use rs_load.table() for importing data sets from text based formats"
}
