% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_loadcsvdata.R
\name{rf_loadtxtdata}
\alias{rf_loadtxtdata}
\title{Read in reffinder data files with auto delimiter detection}
\usage{
rf_loadtxtdata(filepath = "")
}
\arguments{
\item{filepath}{An optional file path to a data file. The function will try to find all files with same extension to load.}
}
\value{
A list or list of lists of input data tables
}
\description{
Read in reffinder data files with auto delimiter detection
}
\examples{

\dontrun{
rf_loadtxtdata()
}



}
