% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_miRNA.R
\docType{data}
\name{rs_miRNA}
\alias{rs_miRNA}
\title{miRNA expression data from ovarian cancer patients}
\format{
A tibble containing miRNA Cp values using miRNA ID as columns and samples as rows
}
\usage{
data(rs_miRNA)
}
\description{
miRNA expression data from ovarian cancer patients
}
\examples{


data(rs_miRNA)

\dontrun{
rs_reffinder(rs_miRNA)
}
}
\keyword{datasets}
