% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswizard.R
\name{rs_wizard}
\alias{rs_wizard}
\title{Perform stability analysis through interactive wizard}
\usage{
rs_wizard()
}
\value{
Creates graphs and result export to
}
\description{
Perform stability analysis through interactive wizard
}
\examples{

\dontrun{
rs_wizard()
}
}
