% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsdialogs.R
\name{rsdialog}
\alias{rsdialog}
\title{Dialog window to select a folder for output files}
\usage{
rsdialog()
}
\value{
A list of string representing answer given in the dialog
1) Was the ok button pressed? - TRUE/FALSE
2) Output directory selected
3) Graphtype selected - individual/multi
4) Sorting of targets on graph x-axis - reffinder/genorm/normfinder/bestkeeper/deltact/targets
4) File type selected for graph output - png/tiff/jpeg/none
5) File type for data output - xlsx/ods/csv/tsv/txt
6) One or more input data file paths
}
\description{
Dialog window to select a folder for output files
}
\examples{

\dontrun{
dialog.answers <- rsdialog()
}


}
