% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsoutdirselect.R
\name{rsexltypeselect}
\alias{rsexltypeselect}
\title{Dialog window to select an excel file type}
\usage{
rsexltypeselect()
}
\value{
A string representing the selected type
}
\description{
Dialog window to select an excel file type
}
\examples{

\dontrun{
path <- rsoutdirselect()
}


}
