% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_alorithms.R
\name{rsfinder}
\alias{rsfinder}
\title{Calculate reffinder stability tables}
\usage{
rsfinder(expression)
}
\arguments{
\item{expression}{A matrix, data frame or tibble with columns for each targets and rows for each samples.}
}
\value{
A list of two data frame tables containing ranked stabilities and stability values from Normfinder, geNorm, BestKeeper and the reffinders comprehensive ranking
}
\description{
Uses Normfinder, geNorm, BestKeeper and delta-Ct methods to calculate ranked stabilities and the comprehensive ranking from reffinder
}
\examples{
set.seed(100)
ct_vals <- matrix(rnorm(5*20, mean = 25), ncol = 5, nrow = 20)
dimnames(ct_vals)[[2]] <-  c("gene1", "gene2", "gene3", "gene4", "gene5")

\dontrun{
rffndr <- rsfinder(ct_vals)
rffndr

}

}
\references{
F Xie, P Xiao, D Chen, L Xu, B Zhang. 2012. miRDeepFinder: a miRNA analysis tool for deep sequencing of plant small RNAs. Plant molecular biology 80 (1), 75-84.
}
