% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsdialogs.R
\name{rsgraphtypeselect}
\alias{rsgraphtypeselect}
\title{Dialog window to select a graph type}
\usage{
rsgraphtypeselect()
}
\value{
A string representing the selected type
}
\description{
Dialog window to select a graph type
}
\examples{

\dontrun{
path <- rsoutdirselect()
}


}
