% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rforderbystability.R
\name{rforderbystability}
\alias{rforderbystability}
\title{Orders a expression stability data frame by stability value}
\usage{
rforderbystability(df)
}
\arguments{
\item{df}{An expression stability matrix, tibble or data frame, should have expression stabilities in second column}
}
\value{
An ordered data frame
}
\description{
Orders a expression stability data frame by stability value
}
\examples{
ct_vals <- matrix(rnorm(5*20), ncol = 5, nrow = 20)
dimnames(ct_vals)[[2]] <-  c("gene1", "gene2", "gene3", "gene4", "gene5")

dct <- rf_deltaCt(ct_vals)

dct <- rforderbystability(dct)
dct

}
