% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsdialogs.R
\name{rsoutdirselect}
\alias{rsoutdirselect}
\title{Dialog window to select a folder for output files}
\usage{
rsoutdirselect()
}
\value{
A string representing a path to a selected output folder
}
\description{
Dialog window to select a folder for output files
}
\examples{

\dontrun{
path <- rsoutdirselect()
}


}
